requirements.txt
websocket-client
pandas
requests
scipy
numpy
python-binance